Randomly place interactive floorkey
Randomly places a floorkey in one of the
rooms. The user can click or tap on it to
pick it up
