+++ SwipeRogueUnity/Assets/Prefabs/FloorKey.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
 guid: c69666bdd4ede440fa42035a531dbbf2
-timeCreated: 1488944556
+timeCreated: 1489026304
 licenseType: Free
 NativeFormatImporter:
   userData:
