+++ SwipeRogueUnity/Assets/Scripts/FloorKey.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: e34f15ddedf624cc3ba80a6a0ddea1b3
+timeCreated: 1489025678
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
